Modified HUD and Camera
HUD is now 3 separate pieces (this also fixed some fill bugs with the
shield bar)

Camera is no longer a child of the player and will no longer cause the
player to crash into walls when it collides with a wall.

Camera now automatically looks for the player and camera target instead
of needing to manually identify them for each scene.

Scripts now look for tags instead of object names to avoid problems with
dynamically making objects if used.