Turret Variables and Targeting
The Turret variables bulletVel and fireRate are now public variables,
defaulted in the prefabs to their previous settings with the enums.
Remember the velocity is faster the higher bulletVel is, but the rate of
fire is faster when fireRate is lower.

Turrets no longer constantly calculate the distance between themselves
and the player. Instead, a TurretTrigger prefab toggles on/off all
turrets in its public turretList variable upon being triggered by the
player.