Massive Optimization
Removed Raycasters on all non-interactive canvases (First Person HUD,
3RD Person HUD, Bullet Brackets).

The editor still does not maintain a constant 60FPS but in the first
room of the end area it can maintain about 50FPS no problem.

When built there are no longer any performance issues with any of the
rooms.

More stress tests are required, but we should be able to get a fairly
high bullet density now.