PowerUpTurret Framework
The essential components of the PowerUp system on my end are finished.
PowerUpTurrets are currently composed of spheres (contrasting with the
cubic BulletTurrets), which create spherical fields when activated by
the TurretTriggers. The player may shoot the PowerUpTurrets, depleting
their health and eventually killing them. On death they spawn a
PowerUpObj which the player can pick up. Upon being picked up, a state
in the PlayerController will change. The Turret inheritance hierarchy
has been altered a bit. I'm not completely sure whether this will break
any in-game turrets at the moment, but fixing would not be hard.