Player Can Now Use PowerUps
Once the player has killed a PowerUp Turret and acquired the PowerUp
item, the PowerUp may be used by hitting the 'R' button. Instead of the
PlayerController adding and removing script components for the PowerUps
to and from the player at runtime, they are all on the Player prefab at
all times, 99.99% of the time disabled, only enabled to briefly activate
the PowerUp.