Bullet LifeTime and Size
The life time of the bullets of a particular turret is now a public,
adjustable variable for level design.

I reverted the Bullet prefab to its original size, renamed it
"InsideBullet," and create a new prefab called "OutsideBullet" which is
twice the size. No turrets should be broken, but all ObjectPoolers will
use InsideBullets by default as they will be used far more often than
OutsideBullets.