Sensitivity Slider Functionality + Bug Fixes
The mouse sensitivity slider now has functionality.

Fixed issue that would allow the player to shoot and take other actions
while the game is paused.

Removed a handful of Debug statements for performance.

Fixed Issue with colliders not knowing if one of their array slots was
null and attempting to access it.