Removed Old Mesh Filter
Fixed the console spewing garbage.

Added new tag to player model for easy referencing later.

Cleaned up some of the warnings from the compiler.

Added a set bomb charge function to the bomb controller and altered the
commented out code in the reload checkpoint function to use it when
needed.