Updated Bomb
Bomb is once again a child of the player that now will stop being a
child once armed.

Once armed it will add a rigidbody component to itself. DO NOT ADD A
RIGIDBODY TO IT ANY OTHER WAY. IF YOU DO IT WILL NOT BEHAVE PROPERLY.

Moved the bomb gauge from the player to the bomb itself.

Removed extraneous bomb objects from scenes.