Beginnings of Bomb Functionality
The player can now shoot a projectile that will eventually become the
bomb with the F key, though that button is not hooked up to the
InputManager yet. The bomb's velocity and deceleration are determined by
how much of the bomb charge is used. Currently, it defaults to using up
how ever much of the gauge is filled.