Fixed Reactor Boss
Added a check to the reactor boss's bomb detach script. Now the player's
bomb must meet the required number (Currently 100).

I noticed that nobody bothered to copy the ShatterWhenHit script before
changing it. It broke the boss battle and I had to recreate the script
from memory. It's working again.