Added Endings
If the bomb timer reaches 0, the dead ending will load. If the player
hits the escape trigger, the alive ending will load.

Each ending will play an audio clip on a delay. once the clip has ended,
it will then load the main menu.

Both ending scenes have a delay when starting the clip and a delay for
the end of the clip.  The current end delay is 0.