Officially Added Core Visuals to Master
I had to copy and paste backup versions of the scripts that I changed
because Git completely got rid of my changes in them when I merged.

There is now a deactivated turret that I will be using for future
testing of this room that will not affect anything as long as it remains
untouched.