Added Particle Effects (after Git issues)
-There is now a visual representation of the player being EMP'ed. The
thrusters turn off completely, and a cloud is generated in the
center-back of the ship which generates static-like lightning particles.

-Checkpoints now have particle systems in the towers on either side that
are a color which corresponds to whether the player has gone through
them yet or not.

-There was an issue where Github attributed almost 200 changes to
animation files to me despite them not being made by me, and that issue
was keeping Git from being able to push or sync. Luckily, that issue has
been fixed.