Bomb and Shield Updates
-The bomb no longer slows down in mid-air, maintaining a constant
velocity until it explodes or hits another object.

-There is now an adjustable 0.5 second cushion added on to the timer for
the bomb's explosion to account for the removal of the requirement for
the bomb to stop moving in mid-air before the timer even starts.

-The shield now, in addition to restricting the number of bullet
absorption particle effects, also restricts the number of bullets that
can even be absorbed in the first place. This at least subtly reduces
the ability of players to crash the game with PulseTurrets or fully
charge the bomb in a single moment with one.