Made Tag for Checkpoint Objects
I've made a tag for checkpoint objects for a specific purpose. Pushing
to make sure there are no merge issues with the TagManager on my end.
Also going to be deleting the extra models placed on the checkpoints
because they don't fit with the visuals of the actual object and are
unnecessary.