The Giant Death Lasers can once again kill the player.
At some point a Player layer must have been added and assigned to the player object. The GDLs were only looking at the Default and Environment layers, so they never "saw" the player. Adding the Player layer to the mask of layers the GDLs can look at fixed the problem.